// React/app/GerAdv_TS/TipoValorProcesso/Interfaces/interface.TipoValorProcesso.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface ITipoValorProcesso {
// 202501251
    id: number;
 
	descricao: string,
}

export interface ITipoValorProcessoFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface ITipoValorProcessoIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}